Moved loader and saver into a common namespace.
